refactor: move alignment file i/o functions to align_util.py
